More clang warning clean-ups
The short => int change is a UB issue, the second argument to va_start()
must be of a type that does not undergo promotion. See
https://wiki.sei.cmu.edu/confluence/display/cplusplus/EXP58-CPP.+Pass+an+object+of+the+correct+type+to+va_start

+++ src/ioc/db/dbUnitTest.c
@@ -111,7 +111,7 @@ long testdbVPutField(const char* pv, short dbrType, va_list ap)
     return dbPutField(&addr, dbrType, pod.bytes, 1);
 }
 
-void testdbPutFieldOk(const char* pv, short dbrType, ...)
+void testdbPutFieldOk(const char* pv, int dbrType, ...)
 {
     long ret;
     va_list ap;
@@ -123,7 +123,7 @@ void testdbPutFieldOk(const char* pv, short dbrType, ...)
     testOk(ret==0, "dbPutField(%s, %d, ...) == %ld", pv, dbrType, ret);
 }
 
-void testdbPutFieldFail(long status, const char* pv, short dbrType, ...)
+void testdbPutFieldFail(long status, const char* pv, int dbrType, ...)
 {
     long ret;
     va_list ap;
@@ -138,7 +138,7 @@ void testdbPutFieldFail(long status, const char* pv, short dbrType, ...)
         testFail("dbPutField(\"%s\", %d, ...) != %ld (%ld)", pv, dbrType, status, ret);
 }
 
-void testdbGetFieldEqual(const char* pv, short dbrType, ...)
+void testdbGetFieldEqual(const char* pv, int dbrType, ...)
 {
     va_list ap;
